+++ src/main/java/com/ninecookies/wiremock/extensions/CallbackConfiguration.java
@@ -10,6 +10,9 @@ import com.amazon.sqs.javamessaging.SQSConnection;
 import com.amazon.sqs.javamessaging.SQSConnectionFactory;
 import com.amazonaws.auth.DefaultAWSCredentialsProviderChain;
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
+import com.amazonaws.services.sns.AmazonSNS;
+import com.amazonaws.services.sns.AmazonSNSClientBuilder;
+import com.amazonaws.services.sqs.AmazonSQS;
 import com.amazonaws.services.sqs.AmazonSQSClientBuilder;
 import com.ninecookies.wiremock.extensions.util.Strings;
 
@@ -22,8 +25,10 @@ import com.ninecookies.wiremock.extensions.util.Strings;
  * <li>{@code RETRY_BACKOFF} default 5_000
  * <li>{@code MAX_RETRIES} default 0 (means disabled)
  * <li>{@code AWS_REGION} the AWS region for SQS messaging (default empty means SQS messaging disabled).
- * <li>{@code AWS_SQS_ENDPOINT} the SQS endpoint to use for testing with localstack or elasticmq (default empty means
- * AWS SQS messaging is used).
+ * <li>{@code AWS_SQS_ENDPOINT} the SQS endpoint to use for testing with localstack (default empty means
+ * AWS messaging is used).
+ * <li>{@code AWS_SNS_ENDPOINT} the SNS endpoint to use for testing with localstack (default empty means
+ * AWS messaging is used).
  * </ul>
  *
  * @author M.Scheepers
@@ -42,8 +47,8 @@ public class CallbackConfiguration {
     private int retryBackoff;
     private int maxRetries;
     private String region;
-    private String endpoint;
-    private AmazonSQSClientBuilder clientBuilder;
+    private AmazonSQSClientBuilder sqsClientBuilder;
+    private AmazonSNSClientBuilder snsClientBuilder;
     private SQSConnectionFactory connectionFactory;
 
     private CallbackConfiguration() {
@@ -55,21 +60,33 @@ public class CallbackConfiguration {
         retryBackoff = parseEnvironmentSetting("RETRY_BACKOFF", DEFAULT_RETRY_BACKOFF);
         maxRetries = parseEnvironmentSetting("MAX_RETRIES", DEFAULT_MAX_RETRIES);
         region = System.getenv("AWS_REGION");
-        endpoint = System.getenv("AWS_SQS_ENDPOINT");
 
         if (!Strings.isNullOrEmpty(region)) {
-            clientBuilder = AmazonSQSClientBuilder.standard()
+            sqsClientBuilder = AmazonSQSClientBuilder.standard()
                     .withCredentials(new DefaultAWSCredentialsProviderChain());
-            if (Strings.isNullOrEmpty(endpoint)) {
+            String sqsEndpoint = System.getenv("AWS_SQS_ENDPOINT");
+            if (Strings.isNullOrEmpty(sqsEndpoint)) {
                 LOG.debug("amazonSQS with region '{}'", region);
-                clientBuilder.withRegion(region);
+                sqsClientBuilder.withRegion(region);
             } else {
-                LOG.warn("amazonSQS with region '{}' and endpoint '{}'", region, endpoint);
-                clientBuilder.setEndpointConfiguration(new EndpointConfiguration(endpoint, region));
+                LOG.warn("amazonSQS with region '{}' and endpoint '{}'", region, sqsEndpoint);
+                sqsClientBuilder.setEndpointConfiguration(new EndpointConfiguration(sqsEndpoint, region));
+            }
+
+            snsClientBuilder = AmazonSNSClientBuilder.standard()
+                    .withCredentials(new DefaultAWSCredentialsProviderChain());
+            String snsEndpoint = System.getenv("AWS_SNS_ENDPOINT");
+            if (Strings.isNullOrEmpty(snsEndpoint)) {
+                LOG.debug("amazonSNS with region '{}'", region);
+                snsClientBuilder.withRegion(region);
+            } else {
+                LOG.warn("amazonSNS with region '{}' and endpoint '{}'", region, snsEndpoint);
+                snsClientBuilder.setEndpointConfiguration(new EndpointConfiguration(snsEndpoint, region));
             }
         } else {
-            LOG.info("AWS SQS messaging callbacks disabled");
-            clientBuilder = null;
+            LOG.info("AWS SNS/SQS messaging callbacks disabled");
+            sqsClientBuilder = null;
+            snsClientBuilder = null;
         }
     }
 
@@ -114,36 +131,62 @@ public class CallbackConfiguration {
     }
 
     /**
-     * Indicates whether SQS messaging is enabled.
+     * Indicates whether SNS/SQS messaging is enabled.
      *
-     * @return {@code true} if SQS callback messaging can be used; otherwise {@code false}.
+     * @return {@code true} if SNS/SQS callback messaging can be used; otherwise {@code false}.
      */
-    public boolean isSqsMessagingEnabled() {
-        return clientBuilder != null;
+    public boolean isMessagingEnabled() {
+        return !Strings.isNullOrEmpty(region);
+    }
+
+    /**
+     * Creates a new Amazon SNS client instance.
+     *
+     * @return a new {@link AmazonSNS} ready to use or {@code null} if {@link #isSnsMessagingEnabled()} is
+     *         {@code false}.
+     */
+    public AmazonSNS createSnsClient() {
+        if (!isMessagingEnabled()) {
+            return null;
+        }
+        return snsClientBuilder.build();
+    }
+
+    /**
+     * Creates a new Amazon SQS client instance.
+     *
+     * @return a new {@link AmazonSQS} ready to use or {@code null} if {@link #isSnsMessagingEnabled()} is
+     *         {@code false}.
+     */
+    public AmazonSQS createSqsClient() {
+        if (!isMessagingEnabled()) {
+            return null;
+        }
+        return sqsClientBuilder.build();
     }
 
     /**
      * Creates a new connection factory instance.
      *
-     * @return a new {@link SQSConnectionFactory} ready to use or {@code null} if {@link #isSqsMessagingEnabled()} is
+     * @return a new {@link SQSConnectionFactory} ready to use or {@code null} if {@link #isMessagingEnabled()} is
      *         {@code false}.
      */
     public SQSConnectionFactory createConnectionFactory() {
-        if (!isSqsMessagingEnabled()) {
+        if (!isMessagingEnabled()) {
             return null;
         }
-        return new SQSConnectionFactory(new ProviderConfiguration(), clientBuilder.build());
+        return new SQSConnectionFactory(new ProviderConfiguration(), sqsClientBuilder.build());
     }
 
     /**
      * Creates a new SQS connection ready to use.
      *
-     * @return a new {@link SQSConnection} ready to use or {@code null} if {@link #isSqsMessagingEnabled()} is
+     * @return a new {@link SQSConnection} ready to use or {@code null} if {@link #isMessagingEnabled()} is
      *         {@code false}.
      * @throws JMSException if a connection couldn't be established.
      */
     public SQSConnection createConnection() throws JMSException {
-        if (!isSqsMessagingEnabled()) {
+        if (!isMessagingEnabled()) {
             return null;
         }
         if (connectionFactory == null) {