+++ src/main/java/com/ninecookies/wiremock/extensions/HttpCallbackHandler.java
@@ -3,10 +3,7 @@ package com.ninecookies.wiremock.extensions;
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpHost;
@@ -29,25 +26,19 @@ import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.protocol.HttpContext;
 import org.apache.http.util.EntityUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import com.github.tomakehurst.wiremock.common.Json;
+import com.ninecookies.wiremock.extensions.HttpCallbackHandler.HttpCallback;
 import com.ninecookies.wiremock.extensions.api.Authentication;
 
 /**
  * Implements {@link Runnable} and uses {@link HttpPost} in combination with {@link HttpEntity} and
  * {@link HttpContext} to emit a POST request according to the referenced callback definition.
  */
-public class HttpCallbackHandler implements Runnable {
+public class HttpCallbackHandler extends AbstractCallbackHandler<HttpCallback> {
 
-    public static class HttpCallback {
+    public static class HttpCallback extends AbstractCallbackHandler.AbstractCallback {
         /**
-         * The period of time in milliseconds to wait before the callback {@link #data} is POSTed to the {@link #url}.
-         */
-        public long delay;
-        /**
-         * The destination URL to POST the {@link #data} to after {@link #delay} has elapsed.
+         * The destination URL to POST the data to after delay has elapsed.
          */
         public String url;
         /**
@@ -58,13 +49,8 @@ public class HttpCallbackHandler implements Runnable {
          * The request id to use for the callback.
          */
         public String traceId;
-        /**
-         * The object representing arbitrary callback data.
-         */
-        public Object data;
     }
 
-    private static final Logger LOG = LoggerFactory.getLogger(HttpCallbackHandler.class);
     private static final String RPS_TRACEID_HEADER = "X-Rps-TraceId";
     private static final RequestConfig REQUEST_CONFIG = RequestConfig.custom()
             .setSocketTimeout(2_000)
@@ -72,18 +58,14 @@ public class HttpCallbackHandler implements Runnable {
             .setConnectionRequestTimeout(5_000)
             .build();
 
-    private ScheduledExecutorService executor;
-    private int invocation;
-    private int maxRetries;
-    private int retryBackoff;
-    private File callbackFile;
+    private HttpCallbackHandler(ScheduledExecutorService executor, File callbackFile) {
+        super(executor, callbackFile, HttpCallback.class);
+    }
 
     @Override
-    public void run() {
-        LOG.debug("CallbackHandler.run()");
-        boolean delete = false;
+    public void handle(HttpCallback callback) throws CallbackException {
+        getLog().debug("CallbackHandler.run()");
         try {
-            HttpCallback callback = readCallback();
             URI uri = URI.create(callback.url);
             HttpContext context = createHttpContext(uri, callback.authentication);
             HttpEntity content = new StringEntity((String) callback.data, ContentType.APPLICATION_JSON);
@@ -97,67 +79,22 @@ public class HttpCallbackHandler implements Runnable {
                 int status = response.getStatusLine().getStatusCode();
                 if (status >= 200 && status < 300) {
                     // in case of success, just print the status line
-                    LOG.info("post to '{}' succeeded: response: {}", uri, response.getStatusLine());
-                    delete = true;
+                    getLog().info("post to '{}' succeeded: response: {}", uri, response.getStatusLine());
                 } else {
-                    delete = rescheduleIfApplicable();
-                    if (delete) {
-                        String retryInfo = "";
-                        if (maxRetries > 0) {
-                            retryInfo = " after " + maxRetries + " attempts";
-                        }
-                        LOG.warn("post to '{}' failed{}: response: {}\n{}",
-                                uri, retryInfo, response.getStatusLine(), readEntity(response.getEntity()));
-                    } else {
-                        LOG.info("post to '{}' will be retried : response: {}\n{}",
-                                uri, response.getStatusLine(), readEntity(response.getEntity()));
-                    }
+                    throw new RetryCallbackException(String.format(
+                            "post to '%s' failed: response: %s\n%s",
+                            uri, response.getStatusLine(), readEntity(response.getEntity())));
                 }
+            } catch (RetryCallbackException e) {
+                throw e;
             } catch (Exception e) {
-                // in failure case print request body and exception
-                delete = rescheduleIfApplicable();
-                if (delete) {
-                    String retryInfo = "";
-                    if (invocation > 1) {
-                        retryInfo = " after " + invocation + " attempts";
-                    }
-                    LOG.error("post to '{}' errored{}\ncontent {}\n{}", uri, retryInfo, content,
-                            readEntity(content),
-                            e);
-                } else {
-                    LOG.warn("post to '{}' will be retried\ncontent {}\n{}", uri, content, readEntity(content), e);
-                }
-
+                throw new RetryCallbackException(String.format("post to '%s' errored\ncontent %s",
+                        uri, readEntity(content)), e);
             }
+        } catch (RetryCallbackException e) {
+            throw e;
         } catch (Exception e) {
-            delete = true;
-            LOG.error("unable to create http post", e);
-        } finally {
-            if (delete) {
-                try {
-                    Files.deleteIfExists(callbackFile.toPath());
-                } catch (IOException e) {
-                    LOG.error("unable to delete callback definition file", e);
-                }
-            }
-        }
-    }
-
-    private boolean rescheduleIfApplicable() {
-        invocation++;
-        if (invocation <= maxRetries) {
-            executor.schedule(this, retryBackoff * invocation, TimeUnit.MILLISECONDS);
-            return false;
-        }
-        return true;
-    }
-
-    private HttpCallback readCallback() {
-        try {
-            String jsonContent = new String(Files.readAllBytes(callbackFile.toPath()), StandardCharsets.UTF_8);
-            return Json.read(jsonContent, HttpCallback.class);
-        } catch (IOException e) {
-            throw new IllegalStateException("Unable to read callback content from file system", e);
+            throw new CallbackException("unable to create http post", e);
         }
     }
 
@@ -197,13 +134,7 @@ public class HttpCallbackHandler implements Runnable {
         return context;
     }
 
-    public static Runnable of(ScheduledExecutorService executor, int maxRetries, int retryBackoff,
-            File callbackFile) {
-        HttpCallbackHandler result = new HttpCallbackHandler();
-        result.maxRetries = maxRetries;
-        result.retryBackoff = retryBackoff;
-        result.executor = executor;
-        result.callbackFile = callbackFile;
-        return result;
+    public static Runnable of(ScheduledExecutorService executor, File callbackFile) {
+        return new HttpCallbackHandler(executor, callbackFile);
     }
 }